Update 80-20-rail.kcl
Changing the 80/20 rail sample file to only include one rail. Having 4 rails in the base file had undesired impacts on Text-to-cad output. Prompting "one rail of 80/20" would generate one rail, but prompting "a rail of 80/20" would create four. "Two rails of 80/20" returned two rails with different heights.

Changing the base plane of the extrusion so that the rails lay flat on the ground rather than stand on end to avoid confusion when prompting. The height of the rail is the height of the profile. The length of the rail is the overall length.

Changing the profile to be a single sketch and adding detail so the model looks better.

Having some trouble with fillets so leaving them out for now